+++ desktop/app/hooks/use-active-projects.ts
