+++ desktop/app/hooks/use-agent-activity.ts
