+++ desktop/app/hooks/use-agent-chat-log.ts
