+++ desktop/app/routes/_layout.chat.tsx
@@ -6,11 +6,11 @@ import {
   type AgentId,
   type MainAgentId,
   useAgentChatLog,
-} from "@/lib/use-agent-chat-log";
-import { useAgentStatuses } from "@/lib/use-agent-statuses";
-import { COMRADES, type ComradeId } from "@/lib/use-comrade-status";
-import { useContextUsage } from "@/lib/use-context-usage";
-import { type InboxLogRecord, useInboxLog } from "@/lib/use-inbox-log";
+} from "@/hooks/use-agent-chat-log";
+import { useAgentStatuses } from "@/hooks/use-agent-statuses";
+import { COMRADES, type ComradeId } from "@/constants/comrade-config";
+import { useContextUsage } from "@/hooks/use-context-usage";
+import { type InboxLogRecord, useInboxLog } from "@/hooks/use-inbox-log";
 
 const AGENTS: MainAgentId[] = ["noctis", "lunafreya"];
